api/requirements.txt
obd==0.7.3
python-socketio==5.13.0
uvicorn==0.34.2